requirements.txt
kaleido==0.1.0post1
pandas
plotly
umap-learn
matplotlib